+++ src/main/java/com/github/kmu_shell_we/domain/user/repository/UserRepository.java
@@ -1,12 +1,10 @@
 package com.github.kmu_shell_we.domain.user.repository;
 
-import com.github.kmu_shell_we.domain.user.schema.User;
+import com.github.kmu_shell_we.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
 import java.util.UUID;
 
 public interface UserRepository extends JpaRepository<User, UUID> {
 
-    Optional<User> findByEmail(String email);
 }
